requirements.txt
streamlit >= 1.32
python-dotenv
flask
openai
